Protect against deleting pickles that are still needed

+++ nff/train/parallel.py
@@ -90,8 +90,11 @@ def del_grad(rank,
         folder = os.path.join(weight_path, str(rank))
         for file in os.listdir(folder):
             if file.startswith("grad") and file.endswith("pickle"):
-                file_path = os.path.join(folder, file)
-                os.remove(file_path)
+                this_batch= int(file.split("_")[2])
+                # remove things that are more than 10 batches old
+                if batch_num - this_batch >= 10:
+                    file_path = os.path.join(folder, file)
+                    os.remove(file_path)
 
 
 def update_optim(optimizer,
